Basic rulekeeper behavior
The rulekeeper's basic functions have been added; it can navigate the new loaded scenes and idle in certain areas of the map at waypoints. If the player breaks a rule, the rulekeeper will chase down the player.

There is a small bug with the interaction items no longer being detected. This function was working previously and needs to be investigated how the change came to be.