Full conversion to newmesh
Don't worry! The large number of file changes is changing the read/write value on a bunch of assets.

The new room has been updated to work with the navmesh and the rulekeeper. I have decided NOT to implement rebaking of the navmesh into this build due to the high volume of new assets. Trying to rebake the scene in-game with the new models causes the navmesh to glitch out along with the Rulekeeper. It is workable, but not for this sprint.

Added the object shader to all the interactable objects, and works as intended. Need to work with Moore on applying it ONLY to the target, how to remove the shader from unintended targets.