IInteractable Init commit
1) Refactored the interaction class into the IInteractable interface. Those child scripts are also responsible for inheriting monobehaviour again.
2) Added a new RemoveHighlight function to remove the shader effect once the crosshair is no longer on top of the item.
3) Removed a bunch of un-checked scenes from the buildprofile.
4) IInteractable is no longer responsible for tracking a press-and-hold function.

Still need to implement the press-and-hold function from the unity input system.

NOTE: ItemInstance does not employ a general highlight for all its children but it should for. Alternatively, if the highlight shader can be applied to all objects the same way, it could be applied in the Interface instead.