Finished adding Highlight
Responsive SFX are now fully in the game, the player is informed if they are able to interact with an object and if they were successful. The highlight works now for single targets, highlighting only the Target, and removing the highlight from previous Targets. The highlight shader itself is still dealing with some bugs described in the pull request.