Navmesh updates properly
The sceneloader was updated to contain a navmesh that is rebuilt whenever the new scene is loaded. There are a LOT of imported objects, and they have a 'read/write' property turned off that interferes with the navMesh. Easy to solve, and explained in the sceneLoader documentation.

Also fixed some edge cases where the loading screen wouldn't appear for the entire loading process.