SceneLoader Finalized
Removed the legacy SceneLoader script completely, it is no longer attached to the most recent architecture. The buttons in the elevator are handled by the ElevatorScript, being enabled / disabled based on Task completion and what the next available level is.

TODO: Need to implement the papers being added to the Elevator script. Need to move the TaskComplete event somewhere, have it be called by the tasks.