Player is properly reset
Fixed a bug where the last known interactable object will try to remove the highlight and throw and error - this is now caught properly in the IInteractable interface.

Player can now be teleported back to the elevator by enabling and disabling isKinematic when moving the player. Otherwise the rigidbody does not allow you to move the object to a new location.

The rulekeeper's attack script properly calls the loadLevel event to send the player back to the start of the level without respawning the papers on the corkboard.

Made a quick fix to the handbook that sends the player back to the mainmenu properly using the SceneLoader.

This was a lot of changes for a single push, but errors made it unreasonable to push a non-functioning build. I will make an effort to prevent this in the future.