Updating the highlight script
All the old items were using the original highlight script, which stored a variable called isHighlighted that would determine if the player could highlight and then pick up the items. I moved this variable into the new highlight script, and moved the distance calculation to a bounding circle trigger instead.