Navmesh deletion glitch fixed
For whatever reason, the navmesh of a level is deleted whenever the current scene is reloaded. If the player moves to a different scene it works completely fine. The fix is to store the NavMeshData when loading the same scene, add it back to NavMeshSurface after it has been deleted, and then use NavMesh.AddNavMeshData() to update the loaded scene.